refactor(api): API 버전 관리 제거 및 로직 통합
- 기존 v1, v2로 분리되어 있던 API 로직을 단일 엔드포인트로 통합함

이번 리팩토링으로 컨트롤러, 서비스 등의 패키지에서 버전 구분을 제거하고, 모든 쿠폰 관련 API를 `/api/coupon` 하위 경로로 일원화했음.
