--- coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/application/v1/CouponServiceV1.java
@@ -1,8 +1,0 @@
-package dev.be.coupon.api.coupon.application.v1;
-
-import dev.be.coupon.api.coupon.application.dto.CouponIssueCommand;
-import dev.be.coupon.api.coupon.application.dto.CouponIssueResult;
-
-public interface CouponServiceV1 {
-    CouponIssueResult issue(final CouponIssueCommand command);
-}
