+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/presentation/CouponController.java
@@ -1,19 +1,18 @@
-package dev.be.coupon.api.coupon.presentation.v1;
+package dev.be.coupon.api.coupon.presentation;
 
 import dev.be.coupon.api.auth.presentation.AuthenticationPrincipal;
 import dev.be.coupon.api.auth.presentation.dto.LoginUser;
-import dev.be.coupon.api.coupon.application.v1.CouponServiceImplV1;
 import dev.be.coupon.api.coupon.application.dto.CouponCreateResult;
 import dev.be.coupon.api.coupon.application.dto.CouponIssueCommand;
-import dev.be.coupon.api.coupon.application.dto.CouponIssueResult;
 import dev.be.coupon.api.coupon.application.dto.CouponUsageCommand;
 import dev.be.coupon.api.coupon.application.dto.CouponUsageResult;
+import dev.be.coupon.api.coupon.application.CouponService;
 import dev.be.coupon.api.coupon.presentation.dto.CouponCreateRequest;
 import dev.be.coupon.api.coupon.presentation.dto.CouponCreateResponse;
 import dev.be.coupon.api.coupon.presentation.dto.CouponIssueRequest;
-import dev.be.coupon.api.coupon.presentation.dto.CouponIssueResponse;
 import dev.be.coupon.api.coupon.presentation.dto.CouponUsageResponse;
 import dev.be.coupon.common.support.response.CommonResponse;
+import dev.be.coupon.domain.coupon.CouponIssueRequestResult;
 import dev.be.coupon.domain.coupon.exception.UnauthorizedAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,17 +25,18 @@ import org.springframework.web.bind.annotation.RestController;
 import java.net.URI;
 import java.util.UUID;
 
-@RequestMapping("/api")
+@RequestMapping("/api/coupon")
 @RestController
-public class CouponV1Controller implements CouponV1ControllerDocs {
+public class CouponController implements CouponControllerDocs {
 
-    private final CouponServiceImplV1 couponServiceImpl;
+    private final CouponService couponService;
 
-    public CouponV1Controller(final CouponServiceImplV1 couponServiceImpl) {
-        this.couponServiceImpl = couponServiceImpl;
+    public CouponController(final CouponService couponService) {
+        this.couponService = couponService;
     }
 
-    @PostMapping(value = "/coupon/")
+    @Override
+    @PostMapping
     public ResponseEntity<CommonResponse<CouponCreateResponse>> create(
             @AuthenticationPrincipal final LoginUser loginUser,
             @RequestBody final CouponCreateRequest request) {
@@ -44,13 +44,14 @@ public class CouponV1Controller implements CouponV1ControllerDocs {
         if (loginUser == null || loginUser.id() == null) {
             throw new UnauthorizedAccessException("로그인된 사용자만 쿠폰을 생성할 수 있습니다.");
         }
-        CouponCreateResult result = couponServiceImpl.create(request.toCommand(loginUser.id()));
+        CouponCreateResult result = couponService.create(request.toCommand(loginUser.id()));
         return ResponseEntity.created(URI.create("/api/coupon/" + result.id()))
                 .body(CommonResponse.success(CouponCreateResponse.from(result)));
     }
 
-    @PostMapping(value = "/v1/coupon/{couponId}/issue")
-    public ResponseEntity<CommonResponse<CouponIssueResponse>> issue(
+    @Override
+    @PostMapping(value = "/v2/coupon/{couponId}/issue")
+    public ResponseEntity<CommonResponse<String>> issue(
             @AuthenticationPrincipal final LoginUser loginUser,
             @PathVariable("couponId") final UUID couponId) {
 
@@ -59,48 +60,43 @@ public class CouponV1Controller implements CouponV1ControllerDocs {
         }
 
         CouponIssueCommand command = new CouponIssueCommand(loginUser.id(), couponId);
-        CouponIssueResult result = couponServiceImpl.issue(command);
-
-        HttpStatus status = resolveStatus(result);
-        return ResponseEntity.status(status)
-                .body(CommonResponse.success(CouponIssueResponse.from(result)));
+        return getCommonResponseResponseEntity(command);
     }
 
-    @PostMapping(value = "/v1/coupon/{couponId}/issue/test")
-    public ResponseEntity<CommonResponse<CouponIssueResponse>> issue(
+    @PostMapping(value = "/v2/coupon/{couponId}/issue/test")
+    public ResponseEntity<CommonResponse<String>> issue(
             @RequestBody final CouponIssueRequest request,
             @PathVariable("couponId") final UUID couponId) {
 
         CouponIssueCommand command = new CouponIssueCommand(request.userId(), couponId);
-        CouponIssueResult result = couponServiceImpl.issue(command);
-
-        HttpStatus status = resolveStatus(result);
-        return ResponseEntity.status(status)
-                .body(CommonResponse.success(CouponIssueResponse.from(result)));
+        return getCommonResponseResponseEntity(command);
     }
 
-    @PostMapping(value = "/coupon/{couponId}/usage")
+    @Override
+    @PostMapping(value = "/{couponId}/usage")
     public ResponseEntity<CommonResponse<CouponUsageResponse>> usage(
             @AuthenticationPrincipal final LoginUser loginUser,
             @PathVariable final UUID couponId) {
 
         if (loginUser == null || loginUser.id() == null) {
-            throw new UnauthorizedAccessException("로그인된 사용자만 쿠폰을 발급받을 수 있습니다.");
+            throw new UnauthorizedAccessException("로그인된 사용자만 쿠폰을 사용할 수 있습니다.");
         }
 
         CouponUsageCommand command = new CouponUsageCommand(loginUser.id(), couponId);
-        CouponUsageResult result = couponServiceImpl.usage(command);
+        CouponUsageResult result = couponService.usage(command);
         return ResponseEntity.ok().body(CommonResponse.success(CouponUsageResponse.from(result)));
     }
 
-    private HttpStatus resolveStatus(final CouponIssueResult result) {
-        if (result.alreadyIssued()) {
-            return HttpStatus.OK;
-        } else if (result.quantityExceeded()) {
-            return HttpStatus.CONFLICT;
-        } else if (!result.used()) {
-            return HttpStatus.CREATED;
-        }
-        return HttpStatus.INTERNAL_SERVER_ERROR;
+    private ResponseEntity<CommonResponse<String>> getCommonResponseResponseEntity(CouponIssueCommand command) {
+        CouponIssueRequestResult result = couponService.issue(command);
+
+        return switch (result) {
+            case SUCCESS -> ResponseEntity.status(HttpStatus.ACCEPTED)
+                    .body(CommonResponse.success("쿠폰 발급 요청이 성공적으로 접수되었습니다."));
+            case SOLD_OUT -> ResponseEntity.status(HttpStatus.OK)
+                    .body(CommonResponse.success("아쉽지만, 쿠폰이 모두 소진되었습니다."));
+            case DUPLICATE -> ResponseEntity.status(HttpStatus.OK)
+                    .body(CommonResponse.success("이미 참여하셨습니다."));
+        };
     }
 }
